moving code to consumer
