+++ producer-rest-core/src/main/scala/za/co/absa/spline/producer/rest/controller/ExecutionPlansController.scala
@@ -180,15 +180,4 @@ class ExecutionPlansController @Autowired()(
   def executionPlan(@RequestBody execPlan: ExecutionPlan): Future[UUID] = repo
     .insertExecutionPlan(execPlan)
     .map(_ => execPlan.id
-
-  @GetMapping(value = Array("/execution-plans/{plan_id}/data-sources"))
-  @ResponseStatus(HttpStatus.OK)
-  def getAffectsByDataSrc(@PathVariable("plan_id") planId: String
-                         @ApiParam(value= "access")
-                         @RequestParam("access")access: Optional[String]
-  ): Future[Array[String]] = {
-    import scala.concurrent.ExecutionContext.Implicits._
-    repo.getDataSources(planId, access)
-  }
-
 }
