+++ producer-services/src/main/scala/za/co/absa/spline/producer/service/repo/ExecutionProducerRepositoryImpl.scala
@@ -166,55 +166,6 @@ class ExecutionProducerRepositoryImpl @Autowired()(db: ArangoDatabaseAsync) exte
       .addQuery(InsertQuery(EdgeDef.Affects, createExecutionAffects(executionPlan, referencedDSes)))
       .buildTx
   }
-
-  override def isDatabaseOk()(implicit ec: ExecutionContext): Future[Boolean] = {
-    try {
-      val anySplineCollectionName = NodeDef.ExecutionPlan.name
-      val futureIsDbOk = db.collection(anySplineCollectionName).exists.toScala.mapTo[Boolean]
-      futureIsDbOk.foreach { isDbOk =>
-        if (!isDbOk)
-          log.error(s"Collection '$anySplineCollectionName' does not exist. Spline database is not initialized properly!")
-      }
-      futureIsDbOk.recover { case _ => false }
-    } catch {
-      case NonFatal(_) => Future.successful(false)
-    }
-  }
-
-  override def getDataSources(value: String, access: Optional[String])(implicit ec: ExecutionContext): Future[Array[String]] ={
-    val readResult = db.queryStream[String](
-      """
-        FOR d in dataSource
-        |FOR dp in depends
-        |FILTER dp._from == @planId AND dp._to == d._id
-        |RETURN d.uri
-        |""".stripMargin
-      .stripMargin,
-    Map("planId" -> "executionPlan".concat("/").concat(value)
-    )
-    )
-
-    val writeResult = db.queryStream[String](
-      """
-        FOR d in dataSource
-        |FOR aff in affects
-        |FILTER aff._from == @planId AND aff._to == d._id
-        |RETURN d.uri
-        |""".stripMargin
-        .stripMargin,
-      Map("planId" -> "executionPlan".concat("/").concat(value)
-      )
-    )
-
-    val totalResult = readResult.zip(writeResult).map{case(listOne,listTwo) => listOne + listTwo}
-
-    if(access.isEmpty)
-      totalResult.map(_.toArray)
-    else if(access.equals("read"))
-      readResult.map(_.toArray)
-    else
-      writeResult.map(_.toArray)
-  }
 }
 
 object ExecutionProducerRepositoryImpl {
